% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LR04_MISboundaries}
\alias{LR04_MISboundaries}
\title{Marine isotope stages (MIS) boundaries.}
\format{An object of class \code{data.frame} with 232 rows and 7 columns.}
\source{
\url{http://www.lorraine-lisiecki.com/LR04_MISboundaries.txt}
}
\usage{
LR04_MISboundaries
}
\description{
From http://www.lorraine-lisiecki.com/LR04_MISboundaries.txt
}
\details{
@format A data frame with 232 rows and 2 variables:
\describe{
  \item{MIS_Boundary}{Marine isotope stage boundary}
  \item{LR04_Age_ka}{x 1000 years ago}
  }
}
\keyword{datasets}
