% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spratt2016}
\alias{spratt2016}
\title{Global Sea Level Reconstruction using Stacked Records from 0-800 ka.}
\format{An object of class \code{data.frame} with 799 rows and 9 columns.}
\source{
\url{https://www.ncdc.noaa.gov/paleo-search/study/19982}
}
\usage{
spratt2016
}
\description{
This is a Late Pleistocene sea level stack based on marine sediment core data (foraminiferal carbonate d18O) as estimated by several different techniques in seven different studies. Obtained from https://www.ncdc.noaa.gov/paleo-search/study/19982 on 28 June 2017
}
\keyword{datasets}
